+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # (Unreleased; add upcoming change notes here)
+
 - iodide notebooks now automatically save to the server
 - fix bug where we showed incorrect revision save times in history viewer,
   also show revision save time down to the second level
+- alpha now less extreme
 
 # 0.3.0 (2019-03-06)